requirements.txt
Django==1.11
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
PyJWT==1.5.0
psycopg2==2.7.1
Faker==0.7.11
requests==2.13.0
celery[redis]==4.0.2
clearbit==0.1.5
mock==2.0.0